Unify and simplify de/serializeROSmessage processing
For most type_id, serialization approach is similar which can be handled in a
common template function, both single and array field. However, for bool and string type,
there are extra check and logic for single but array does not. Although it is weird,
this commit keep the logic intact.

Signed-off-by: jwang <jing.j.wang@intel.com>