Fix rmw_time to Duration_t conversion
The rmw_time_t has 64-bit fields for seconds and nanoseconds. These
values are truncated to the eprosima::fastrtps::Duration_t's fields
which are int32 and uint32. This truncation causes problems when
the seconds value is greater than INT_MAX. For example, when attempting
to create a subscriber with a QoS policy that has a Duration field
(such as lifespan, deadline, or liveliness_leas_duration) where
seconds = INT_MAX and nanoseconds = UINT_MAX.

As a work-around (until Fast-DDS supports 64-bit seconds and nanoseconds)
this change ensures that the seconds value can be represented with a
32-bit integer.

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>