Fix some overwritten errors in rmw_fastrtps.
There are 2 different kinds of errors fixed in here:

1.  When trying to get the topic QoS settings, the
function get_topic_qos() always sets an error when it
fails.  Thus the higher level functions do not also need
to set an error.

2.  When rmw_serialize() calls rmw_serialized_message_resize(),
it may fail deep in rcutils calls.  Those are very low-level
error messages, and not that useful, so instead we clear
those out and report an error at this rmw_fastrtps layer.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>